# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sindra API Documentation",
        "description": "Dokumentasi API untuk sistem tiket & knowledge base",
        "contact": {
            "email": "support@sindra.local"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Main API Server"
        }
    ],
    "paths": {
...
        "/api/knowledge-bases": {
            "get": {
                "tags": [
                    "Knowledge Base"
                ],
                "summary": "Tampilkan semua Knowledge Base",
                "description": "Ambil daftar Knowledge Base dengan filter opsional berdasarkan kategori, status publikasi, atau pencarian judul.",
                "operationId": "47e61c307a4996fd7ddeae803829a766",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Filter berdasarkan kategori",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    },
                    {
                        "name": "is_published",
                        "in": "query",
                        "description": "Filter berdasarkan status publikasi (true/false)",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "example": true
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Cari berdasarkan judul",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Panduan Reset Password"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar Knowledge Base berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Daftar knowledge base berhasil diambil."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/KnowledgeBase"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Knowledge Base"
                ],
                "summary": "Buat Knowledge Base baru",
                "description": "Menambahkan artikel baru ke knowledge base.",
                "operationId": "35d70b37c4c167e0c35fb92c96e2bbe8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "author_id",
                                    "title"
                                ],
                                "properties": {
                                    "author_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Panduan Menggunakan Aplikasi"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Langkah-langkah penggunaan aplikasi dijelaskan di sini."
                                    },
                                    "is_published": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Knowledge Base berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Knowledge base berhasil dibuat."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/KnowledgeBase"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/knowledge-bases/{id}": {
            "get": {
                "tags": [
                    "Knowledge Base"
                ],
                "summary": "Tampilkan detail Knowledge Base",
                "description": "Menampilkan satu Knowledge Base berdasarkan ID.",
                "operationId": "4ef4dc9b23a5b94cb499c429b1928ae0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Knowledge Base",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Knowledge Base ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/KnowledgeBase"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Knowledge Base tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Knowledge Base"
                ],
                "summary": "Perbarui Knowledge Base",
                "description": "Memperbarui artikel knowledge base berdasarkan ID.",
                "operationId": "8430503bffb894aee0340672dc6f120f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Knowledge Base",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "category_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Panduan Update Akun"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Cara memperbarui data akun pengguna."
                                    },
                                    "is_published": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Knowledge Base berhasil diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Knowledge base berhasil diperbarui."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/KnowledgeBase"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Knowledge Base tidak ditemukan"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Knowledge Base"
                ],
                "summary": "Hapus Knowledge Base",
                "description": "Menghapus artikel Knowledge Base berdasarkan ID.",
                "operationId": "872d8cade6d0b30f8adc33a3352a8060",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Knowledge Base",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Knowledge Base berhasil dihapus"
                    },
                    "404": {
                        "description": "Knowledge Base tidak ditemukan"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "KnowledgeBase": {
                "title": "Knowledge Base",
                "description": "Schema untuk entitas Knowledge Base",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "author_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "title": {
                        "type": "string",
                        "example": "Panduan Menggunakan Sistem"
                    },
                    "slug": {
                        "type": "string",
                        "example": "panduan-menggunakan-sistem"
                    },
                    "content": {
                        "type": "string",
                        "example": "Berikut panduan lengkap untuk menggunakan sistem."
                    },
                    "is_published": {
                        "type": "boolean",
                        "example": true
                    },
                    "published_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-15T10:00:00Z"
                    },
                    "category": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 2
                            },
                            "name": {
                                "type": "string",
                                "example": "Panduan Sistem"
                            }
                        },
                        "type": "object",
                        "nullable": true
                    },
                    "author": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "Admin Utama"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}